bootstrap 적용 2

+++ src/main/webapp/member/deleteMemberAction.jsp
@@ -3,13 +3,22 @@
 <%@ page import="vo.*"%>
 <%
 	// Controller 
+	String memberId = request.getParameter("memberId");
+	String memberPw = request.getParameter("memberPw");
+	String deleteAccount = request.getParameter("deleteAccount"); // 탈퇴 동의 checkbox : value값 : true 
+	String redirectUrl="/member/deleteMemberForm.jsp";
+	if(deleteAccount==null||!deleteAccount.equals("true")||memberId==null||memberId.equals("")||memberPw==null||memberPw.equals("")){
+		response.sendRedirect(request.getContextPath()+redirectUrl);
+		return;
+	}
 	Member member = new Member();
-	member.setMemberId(request.getParameter("memberId"));
-	member.setMemberPw(request.getParameter("memberPw"));
+	member.setMemberId(memberId);
+	member.setMemberPw(memberPw);
 	MemberDao memberDao = new MemberDao();
 	int row = memberDao.deleteMember(member);
 	if(row==1){
 		//System.out.println(row+"<--deleteMemberAction");
-		response.sendRedirect(request.getContextPath()+"/logout.jsp");
+		redirectUrl="/member/logout.jsp";
 	}
+	response.sendRedirect(request.getContextPath()+redirectUrl);
 %>
